+++ src/utils/json/Parser.hpp
@@ -31,16 +31,16 @@
 // version 2.2 of Bison.
 
 /**
- ** \file /media/HDD1/DEV/Projet/C-cpp/git/cpp-utils/src/utils/json/Parser.hpp
+ ** \file /home/maxime/Documents/git/cpp-utils/src/utils/json/Parser.hpp
  ** Define the utils::json::parser class.
  */
 
 // C++ LALR(1) parser skeleton written by Akim Demaille.
 
-#ifndef YY_YY_MEDIA_HDD1_DEV_PROJET_C_CPP_GIT_CPP_UTILS_SRC_UTILS_JSON_PARSER_HPP_INCLUDED
-# define YY_YY_MEDIA_HDD1_DEV_PROJET_C_CPP_GIT_CPP_UTILS_SRC_UTILS_JSON_PARSER_HPP_INCLUDED
+#ifndef YY_YY_HOME_MAXIME_DOCUMENTS_GIT_CPP_UTILS_SRC_UTILS_JSON_PARSER_HPP_INCLUDED
+# define YY_YY_HOME_MAXIME_DOCUMENTS_GIT_CPP_UTILS_SRC_UTILS_JSON_PARSER_HPP_INCLUDED
 // //                    "%code requires" blocks.
-#line 18 "/media/HDD1/DEV/Projet/C-cpp/git/cpp-utils/src/utils/json/Parser.yy" // lalr1.cc:372
+#line 18 "/home/maxime/Documents/git/cpp-utils/src/utils/json/Parser.yy" // lalr1.cc:372
 
 
     #include <utility>
@@ -59,7 +59,7 @@
 
 
 
-#line 63 "/media/HDD1/DEV/Projet/C-cpp/git/cpp-utils/src/utils/json/Parser.hpp" // lalr1.cc:372
+#line 63 "/home/maxime/Documents/git/cpp-utils/src/utils/json/Parser.hpp" // lalr1.cc:372
 
 
 # include <vector>
@@ -128,9 +128,9 @@
 # define YYDEBUG 0
 #endif
 
-#line 8 "/media/HDD1/DEV/Projet/C-cpp/git/cpp-utils/src/utils/json/Parser.yy" // lalr1.cc:372
+#line 8 "/home/maxime/Documents/git/cpp-utils/src/utils/json/Parser.yy" // lalr1.cc:372
 namespace utils { namespace json {
-#line 134 "/media/HDD1/DEV/Projet/C-cpp/git/cpp-utils/src/utils/json/Parser.hpp" // lalr1.cc:372
+#line 134 "/home/maxime/Documents/git/cpp-utils/src/utils/json/Parser.hpp" // lalr1.cc:372
 
 
 
@@ -144,7 +144,7 @@ namespace utils { namespace json {
     /// Symbol semantic values.
     union semantic_type
     {
-    #line 49 "/media/HDD1/DEV/Projet/C-cpp/git/cpp-utils/src/utils/json/Parser.yy" // lalr1.cc:372
+    #line 49 "/home/maxime/Documents/git/cpp-utils/src/utils/json/Parser.yy" // lalr1.cc:372
 
     // "Pure" types
     long long int   v_int;
@@ -156,7 +156,7 @@ namespace utils { namespace json {
     utils::json::Array* v_array;
     utils::json::Value* v_value;
 
-#line 160 "/media/HDD1/DEV/Projet/C-cpp/git/cpp-utils/src/utils/json/Parser.hpp" // lalr1.cc:372
+#line 160 "/home/maxime/Documents/git/cpp-utils/src/utils/json/Parser.hpp" // lalr1.cc:372
     };
 #else
     typedef YYSTYPE semantic_type;
@@ -481,11 +481,11 @@ namespace utils { namespace json {
   };
 
 
-#line 8 "/media/HDD1/DEV/Projet/C-cpp/git/cpp-utils/src/utils/json/Parser.yy" // lalr1.cc:372
+#line 8 "/home/maxime/Documents/git/cpp-utils/src/utils/json/Parser.yy" // lalr1.cc:372
 } } // utils::json
-#line 487 "/media/HDD1/DEV/Projet/C-cpp/git/cpp-utils/src/utils/json/Parser.hpp" // lalr1.cc:372
+#line 487 "/home/maxime/Documents/git/cpp-utils/src/utils/json/Parser.hpp" // lalr1.cc:372
 
 
 
 
-#endif // !YY_YY_MEDIA_HDD1_DEV_PROJET_C_CPP_GIT_CPP_UTILS_SRC_UTILS_JSON_PARSER_HPP_INCLUDED
+#endif // !YY_YY_HOME_MAXIME_DOCUMENTS_GIT_CPP_UTILS_SRC_UTILS_JSON_PARSER_HPP_INCLUDED